requirements.txt
sqlalchemy
psycopg2-binary
pandas
tenacity
numpy
langfuse
tweepy
asyncpg
litellm
python-telegram-bot
feedparser
lxml_html_clean
timeago
bs4